CUD functionalities for reading recommendations are now secured. The Create funcitonality for category is also secured

+++ src/main/java/fi/haagahelia/coolreads/controller/ReadingRecommendationController.java
@@ -5,18 +5,17 @@ import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
 import fi.haagahelia.coolreads.dto.AddReadingRecommendationDto;
 
 import org.springframework.web.bind.annotation.PathVariable;
 
+import fi.haagahelia.coolreads.model.AppUser;
 import fi.haagahelia.coolreads.model.Category;
 import fi.haagahelia.coolreads.model.ReadingRecommendation;
 import fi.haagahelia.coolreads.repository.AppUserRepository;
@@ -33,7 +32,6 @@ public class ReadingRecommendationController {
 
 	@Autowired
 	private AppUserRepository userRepository;
-	
 
 	@GetMapping("/")
 	public String renderReccomendationlist(Model model) {
@@ -51,12 +49,19 @@ public class ReadingRecommendationController {
 
 	@PostMapping("/recommendations/add")
 	public String addMessage(
-			@ModelAttribute("readingRecommendationDto") AddReadingRecommendationDto readingRecommendation) {
+			@ModelAttribute("readingRecommendationDto") AddReadingRecommendationDto readingRecommendation,
+			@AuthenticationPrincipal UserDetails userDetails) {
 
 		Category category = categoryRepository.findByName(readingRecommendation.getCategoryName());
-//		ReadingRecommendation newReadingRecommendation = new ReadingRecommendation(readingRecommendation.getTitle(),
-//				readingRecommendation.getLink(), readingRecommendation.getDescription(), category);
-//		readingRepository.save(newReadingRecommendation);
+
+		AppUser user = userRepository.findOneByUsername(userDetails.getUsername()).orElse(null);
+
+		if (user == null)
+			return "redirect:/";
+
+		ReadingRecommendation newReadingRecommendation = new ReadingRecommendation(readingRecommendation.getTitle(),
+				readingRecommendation.getLink(), readingRecommendation.getDescription(), category, user);
+		readingRepository.save(newReadingRecommendation);
 
 		return "redirect:/";
 	}
@@ -77,12 +82,26 @@ public class ReadingRecommendationController {
 
 	@PostMapping("/recommendations/edit/{id}")
 	public String editRecommendation(@PathVariable Long id,
-			@ModelAttribute("readingRecommendationDto") AddReadingRecommendationDto readingRecommendation) {
+			@ModelAttribute("readingRecommendationDto") AddReadingRecommendationDto readingRecommendation,
+			@AuthenticationPrincipal UserDetails userDetails) {
 		Category category = categoryRepository.findByName(readingRecommendation.getCategoryName());
-//		ReadingRecommendation updatedRecommendation = new ReadingRecommendation(readingRecommendation.getTitle(),
-//				readingRecommendation.getLink(), readingRecommendation.getDescription(), category);
-//		updatedRecommendation.setId(id);
-//		readingRepository.save(updatedRecommendation);
+
+		ReadingRecommendation oldRecommendation = readingRepository.findById(id).orElse(null);
+
+		AppUser user = userRepository.findOneByUsername(userDetails.getUsername()).orElse(null);
+
+		if (user == null || oldRecommendation == null)
+			return "redirect:/";
+
+		if (user.getUserId() != oldRecommendation.getAppUser().getUserId())
+			return "redirect:/";
+
+		oldRecommendation.setCategory(category);
+		oldRecommendation.setDescription(readingRecommendation.getDescription());
+		oldRecommendation.setTitle(readingRecommendation.getTitle());
+		oldRecommendation.setLink(readingRecommendation.getLink());
+
+		readingRepository.save(oldRecommendation);
 
 		return "redirect:/";
 	}
@@ -94,7 +113,18 @@ public class ReadingRecommendationController {
 	}
 
 	@PostMapping("/recommendations/delete")
-	public String deleteRecommendation(@RequestParam("id") Long readingRecommendationId) {
+	public String deleteRecommendation(@RequestParam("id") Long readingRecommendationId,
+			@AuthenticationPrincipal UserDetails userDetails) {
+		ReadingRecommendation recommendationToDelete = readingRepository.findById(readingRecommendationId).orElse(null);
+
+		AppUser user = userRepository.findOneByUsername(userDetails.getUsername()).orElse(null);
+
+		if (user == null || recommendationToDelete == null) 
+			return "redirect:/";
+
+		if (user.getUserId() != recommendationToDelete.getAppUser().getUserId())
+			return "redirect:/";
+
 		readingRepository.deleteById(readingRecommendationId);
 		return "redirect:/";
 	}
